requirements.txt
# Core bot framework (Discord + voice)
discord.py>=2.4.0
PyNaCl>=1.5.0          # required for Discord voice encryption

# LLM (Shape.inc via OpenAI-compatible client)
openai>=1.40.0         # OpenAI SDK v1 (used with Shape.inc base_url)

# Env + HTTP + Web (Flask dashboard)
python-dotenv>=1.0.1
requests>=2.31.0
Flask>=3.0.0

# TTS fallback
gTTS>=2.5.0

# (Optional) ElevenLabs official SDK — NOT required by current code (we call REST via requests)
# elevenlabs>=1.0.0
